Fix for the culling/camera issue
Camera was rotating counter to the frustum because I forgot to rotate the world in the opposite direction to the intended camera rotation.

Renamed the sine/cosineThetas to inverseSine/CosineTheta, because they're now inverted.

Removed a ton of comments from main containing old code and test cases because they're no longer necessary.