Fix for drawing edges
There was a problem in the rendering where the first vert had the index 1 but the first vert in the buffer would, of course, be index 0.

Also fixed a problem with the Mario model's verts. Because I deleted the hair cap component, the indices for all the verts after that were wrong. I manually edited the indices to what they should be.